Fix Resolving of SparkSession Table's Metadata Tables
Do to an issue within Spark's Resolution rules we cannot acces a table in the session
catalog with a multipart identifier. Because we also cannot determine whether the underlying
Iceberg table is Hadoop or Hive based we also cannot know the right method for reading the
table. To work around this for now we attempt to first load the table as a HiveTable, if
the table isn't found we fall back and attempt to load it as a Hadoop table.