Core: Fix Replace Table with Changing Partition Spec
Originally when we rebuild the partition spec for V1 tables we automatically
include void transforms for and source column which is not be transformed
identically in the new replacement table. This has issues if the column's
partitioning transform has changed leading to errors because a spec can
only transform a given column a single time. To fix this we instead replace
any transform in the previous spec that was modifying the same source column
as the new spec. Voids are still propagated for source columns which are
untransformed in the new spec.